+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## 1.5.0 - Upcoming
+
+* Deprecate non-iterable command collections
+
 ## 1.4.0 - 2026-05-18
 
 * Add PHP 8.5 support